backend/pyproject.toml
[project]
name = "embeddings-pipeline"
version = "0.1.0"
description = "Pipeline to extract content from Docusaurus-based book websites, generate embeddings using Cohere models, and store them in Qdrant vector database"
authors = [{name = "Developer", email = "developer@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.0.0",
    "qdrant-client>=1.9.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]